hal: Update local cache only when mxr ctl matches
Device switch latency values had increased as there was
a change to update local cached BE info for all devices
if there's any BE update. So we need to update local cache
of only those BE ids whose mixer control matches
that of current BE id when there is a backend change.

Change-Id: I27d8399049319fcb9029e163a58ece2f42f5c33b